requirements.txt
pymupdf
python-doctr[torch]
torch
torchvision

